# core/src/main/resources/swagger_spec.json
{
  "swagger" : "2.0",
  "info" : {
    "version" : "1.2.0",
    "title" : "Klaw Core",
    "contact" : {
      "email" : "info@klaw-project.io"
    }
  },
  "basePath" : "/",
  "paths" : {
...
    "/createTopics" : {
      "post" : {
        "operationId" : "createTopicsRequest",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/TopicRequestModel"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/ApiResponse"
            }
          }
        }
      }
    },
...
    "/getAcls" : {
      "get" : {
        "operationId" : "getAcls",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "topicnamesearch",
          "in" : "query",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/TopicOverview"
            }
          }
        }
      }
    },
...
    "/getConnectorDetails" : {
      "get" : {
        "operationId" : "getConnectorDetails",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "env",
          "in" : "query",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "connectorName",
          "in" : "query",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "object",
              "additionalProperties" : {
                "type" : "string"
              }
            }
          }
        }
      }
    },
...
    "/getConnectors" : {
      "get" : {
        "operationId" : "getTopics",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "env",
          "in" : "query",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "pageNo",
          "in" : "query",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "currentPage",
          "in" : "query",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "connectornamesearch",
          "in" : "query",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "teamName",
          "in" : "query",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "type" : "array",
                "items" : {
                  "$ref" : "#/definitions/KafkaConnectorModel"
                }
              }
            }
          }
        }
      }
    },
    "/getConsumerOffsets" : {
      "get" : {
        "operationId" : "getConsumerOffsets",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "env",
          "in" : "query",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "topicName",
          "in" : "query",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "consumerGroupId",
          "in" : "query",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "type" : "object",
                "additionalProperties" : {
                  "type" : "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/getTopics" : {
      "get" : {
        "operationId" : "getTopics",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "env",
          "in" : "query",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "pageNo",
          "in" : "query",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "currentPage",
          "in" : "query",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "topicnamesearch",
          "in" : "query",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "teamName",
          "in" : "query",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "topicType",
          "in" : "query",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "type" : "array",
                "items" : {
                  "$ref" : "#/definitions/TopicInfo"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "definitions" : {
    "AclInfo" : {
      "type" : "object",
      "properties" : {
        "sequence" : {
          "type" : "string"
        },
        "req_no" : {
          "type" : "string"
        },
        "acl_ip" : {
          "type" : "string"
        },
        "acl_ssl" : {
          "type" : "string"
        },
        "topicname" : {
          "type" : "string"
        },
        "topictype" : {
          "type" : "string"
        },
        "consumergroup" : {
          "type" : "string"
        },
        "environment" : {
          "type" : "string"
        },
        "environmentName" : {
          "type" : "string"
        },
        "teamname" : {
          "type" : "string"
        },
        "teamid" : {
          "type" : "integer",
          "format" : "int32"
        },
        "operation" : {
          "type" : "string"
        },
        "permission" : {
          "type" : "string"
        },
        "transactionalId" : {
          "type" : "string"
        },
        "aclPatternType" : {
          "type" : "string"
        },
        "totalNoPages" : {
          "type" : "string"
        },
        "allPageNos" : {
          "type" : "array",
          "items" : {
            "type" : "string"
          }
        },
        "possibleTeams" : {
          "type" : "array",
          "items" : {
            "type" : "string"
          }
        },
        "currentPage" : {
          "type" : "string"
        },
        "showDeleteAcl" : {
          "type" : "boolean"
        },
        "kafkaFlavorType" : {
          "type" : "string",
          "enum" : [ "APACHE_KAFKA", "AIVEN_FOR_APACHE_KAFKA", "CONFLUENT", "OTHERS" ]
        }
      }
    },
...
    "ApiResponse" : {
      "type" : "object",
      "properties" : {
        "status" : {
          "type" : "string",
          "enum" : [ "CONTINUE", "SWITCHING_PROTOCOLS", "PROCESSING", "CHECKPOINT", "OK", "CREATED", "ACCEPTED", "NON_AUTHORITATIVE_INFORMATION", "NO_CONTENT", "RESET_CONTENT", "PARTIAL_CONTENT", "MULTI_STATUS", "ALREADY_REPORTED", "IM_USED", "MULTIPLE_CHOICES", "MOVED_PERMANENTLY", "FOUND", "MOVED_TEMPORARILY", "SEE_OTHER", "NOT_MODIFIED", "USE_PROXY", "TEMPORARY_REDIRECT", "PERMANENT_REDIRECT", "BAD_REQUEST", "UNAUTHORIZED", "PAYMENT_REQUIRED", "FORBIDDEN", "NOT_FOUND", "METHOD_NOT_ALLOWED", "NOT_ACCEPTABLE", "PROXY_AUTHENTICATION_REQUIRED", "REQUEST_TIMEOUT", "CONFLICT", "GONE", "LENGTH_REQUIRED", "PRECONDITION_FAILED", "PAYLOAD_TOO_LARGE", "REQUEST_ENTITY_TOO_LARGE", "URI_TOO_LONG", "REQUEST_URI_TOO_LONG", "UNSUPPORTED_MEDIA_TYPE", "REQUESTED_RANGE_NOT_SATISFIABLE", "EXPECTATION_FAILED", "I_AM_A_TEAPOT", "INSUFFICIENT_SPACE_ON_RESOURCE", "METHOD_FAILURE", "DESTINATION_LOCKED", "UNPROCESSABLE_ENTITY", "LOCKED", "FAILED_DEPENDENCY", "UPGRADE_REQUIRED", "PRECONDITION_REQUIRED", "TOO_MANY_REQUESTS", "REQUEST_HEADER_FIELDS_TOO_LARGE", "UNAVAILABLE_FOR_LEGAL_REASONS", "INTERNAL_SERVER_ERROR", "NOT_IMPLEMENTED", "BAD_GATEWAY", "SERVICE_UNAVAILABLE", "GATEWAY_TIMEOUT", "HTTP_VERSION_NOT_SUPPORTED", "VARIANT_ALSO_NEGOTIATES", "INSUFFICIENT_STORAGE", "LOOP_DETECTED", "BANDWIDTH_LIMIT_EXCEEDED", "NOT_EXTENDED", "NETWORK_AUTHENTICATION_REQUIRED" ]
        },
        "timestamp" : {
          "type" : "string",
          "format" : "date-time"
        },
        "message" : {
          "type" : "string"
        },
        "debugMessage" : {
          "type" : "string"
        },
        "result" : {
          "type" : "string"
        },
        "data" : {
          "type" : "object"
        }
      }
    },
...
    "KafkaConnectorModel" : {
      "type" : "object",
      "properties" : {
        "sequence" : {
          "type" : "integer",
          "format" : "int32"
        },
        "connectorId" : {
          "type" : "integer",
          "format" : "int32"
        },
        "connectorName" : {
          "type" : "string"
        },
        "connectorConfig" : {
          "type" : "string"
        },
        "environmentName" : {
          "type" : "string"
        },
        "environmentId" : {
          "type" : "string"
        },
        "teamName" : {
          "type" : "string"
        },
        "possibleTeams" : {
          "type" : "array",
          "items" : {
            "type" : "string"
          }
        },
        "allPageNos" : {
          "type" : "array",
          "items" : {
            "type" : "string"
          }
        },
        "totalNoPages" : {
          "type" : "string"
        },
        "currentPage" : {
          "type" : "string"
        },
        "remarks" : {
          "type" : "string"
        },
        "documentation" : {
          "type" : "string"
        },
        "environmentsList" : {
          "type" : "array",
          "items" : {
            "type" : "string"
          }
        },
        "description" : {
          "type" : "string"
        },
        "showEditConnector" : {
          "type" : "boolean"
        },
        "showDeleteConnector" : {
          "type" : "boolean"
        },
        "connectorDeletable" : {
          "type" : "boolean"
        }
      }
    },
...
    "TopicConfigEntry" : {
      "type" : "object",
      "properties" : {
        "configKey" : {
          "type" : "string"
        },
        "configValue" : {
          "type" : "string"
        }
      }
    },
    "TopicHistory" : {
      "type" : "object",
      "properties" : {
        "environmentName" : {
          "type" : "string"
        },
        "teamName" : {
          "type" : "string"
        },
        "requestedBy" : {
          "type" : "string"
        },
        "requestedTime" : {
          "type" : "string"
        },
        "approvedBy" : {
          "type" : "string"
        },
        "approvedTime" : {
          "type" : "string"
        },
        "remarks" : {
          "type" : "string"
        }
      }
    },
    "TopicInfo" : {
      "type" : "object",
      "properties" : {
        "topicid" : {
          "type" : "integer",
          "format" : "int32"
        },
        "sequence" : {
          "type" : "string"
        },
        "totalNoPages" : {
          "type" : "string"
        },
        "currentPage" : {
          "type" : "string"
        },
        "allPageNos" : {
          "type" : "array",
          "items" : {
            "type" : "string"
          }
        },
        "topicName" : {
          "type" : "string"
        },
        "noOfPartitions" : {
          "type" : "integer",
          "format" : "int32"
        },
        "description" : {
          "type" : "string"
        },
        "documentation" : {
          "type" : "string"
        },
        "noOfReplcias" : {
          "type" : "string"
        },
        "teamname" : {
          "type" : "string"
        },
        "cluster" : {
          "type" : "string"
        },
        "clusterId" : {
          "type" : "string"
        },
        "environmentsList" : {
          "type" : "array",
          "items" : {
            "type" : "string"
          }
        },
        "showEditTopic" : {
          "type" : "boolean"
        },
        "showDeleteTopic" : {
          "type" : "boolean"
        },
        "topicDeletable" : {
          "type" : "boolean"
        }
      }
    },
    "TopicOverview" : {
      "type" : "object",
      "properties" : {
        "topicInfoList" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/TopicInfo"
          }
        },
        "aclInfoList" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/AclInfo"
          }
        },
        "prefixedAclInfoList" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/AclInfo"
          }
        },
        "transactionalAclInfoList" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/AclInfo"
          }
        },
        "topicHistoryList" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/TopicHistory"
          }
        },
        "promotionDetails" : {
          "type" : "object",
          "additionalProperties" : {
            "type" : "string"
          }
        },
        "topicExists" : {
          "type" : "boolean"
        },
        "schemaDetails" : {
          "type" : "array",
          "items" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "string"
            }
          }
        },
        "schemaExists" : {
          "type" : "boolean"
        },
        "prefixAclsExists" : {
          "type" : "boolean"
        },
        "txnAclsExists" : {
          "type" : "boolean"
        },
        "topicDocumentation" : {
          "type" : "string"
        },
        "topicIdForDocumentation" : {
          "type" : "integer",
          "format" : "int32"
        }
      }
    },
    "TopicRequestModel" : {
      "type" : "object",
      "properties" : {
        "topicname" : {
          "type" : "string"
        },
        "environment" : {
          "type" : "string"
        },
        "topicpartitions" : {
          "type" : "integer",
          "format" : "int32"
        },
        "teamname" : {
          "type" : "string"
        },
        "remarks" : {
          "type" : "string"
        },
        "description" : {
          "type" : "string"
        },
        "replicationfactor" : {
          "type" : "string"
        },
        "environmentName" : {
          "type" : "string"
        },
        "topicid" : {
          "type" : "integer",
          "format" : "int32"
        },
        "advancedTopicConfigEntries" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/TopicConfigEntry"
          }
        },
        "appname" : {
          "type" : "string"
        },
        "topictype" : {
          "type" : "string"
        },
        "requestor" : {
          "type" : "string"
        },
        "requesttime" : {
          "type" : "string",
          "format" : "date-time"
        },
        "requesttimestring" : {
          "type" : "string"
        },
        "topicstatus" : {
          "type" : "string"
        },
        "approver" : {
          "type" : "string"
        },
        "approvingtime" : {
          "type" : "string",
          "format" : "date-time"
        },
        "sequence" : {
          "type" : "string"
        },
        "username" : {
          "type" : "string"
        },
        "totalNoPages" : {
          "type" : "string"
        },
        "approvingTeamDetails" : {
          "type" : "string"
        },
        "otherParams" : {
          "type" : "string"
        },
        "teamId" : {
          "type" : "integer",
          "format" : "int32"
        },
        "allPageNos" : {
          "type" : "array",
          "items" : {
            "type" : "string"
          }
        },
        "possibleTeams" : {
          "type" : "array",
          "items" : {
            "type" : "string"
          }
        },
        "currentPage" : {
          "type" : "string"
        }
      }
    },
...
  }
}